requirements.txt
flask
flask-cors
python-dotenv
twilio
google-generativeai
